+++ src/main/kotlin/de/andrena/tools/altn8th/settings/ui/components/general/GeneralLinksUi.kt
@@ -1,9 +1,8 @@
 package de.andrena.tools.altn8th.settings.ui.components.general
 
 import com.intellij.openapi.options.ShowSettingsUtil
-import com.intellij.openapi.project.ProjectManager
-import com.intellij.ui.HyperlinkLabel
 import com.intellij.ui.IdeBorderFactory
+import com.intellij.ui.components.ActionLink
 import com.intellij.ui.components.JBTextArea
 import com.intellij.ui.components.panels.VerticalLayout
 import de.andrena.tools.altn8th.internationalization.I18n
@@ -22,24 +21,18 @@ private const val SPACING = 8
 
 internal class GeneralLinksUi : Ui {
     private val prefixSettingsLink =
-        HyperlinkLabel(I18n.lazyMessage("altn8.pattern.prefixes").get()).apply {
-            addHyperlinkListener {
-                routeToSettingPage(PrefixConfigurable::class.java)
-            }
+        ActionLink(I18n.lazyMessage("altn8.pattern.prefixes").get()) {
+            routeToSettingPage(PrefixConfigurable::class.java)
         }
 
     private val postfixSettingsLink =
-        HyperlinkLabel(I18n.lazyMessage("altn8.pattern.postfixes").get()).apply {
-            addHyperlinkListener {
-                routeToSettingPage(PostfixConfigurable::class.java)
-            }
+        ActionLink(I18n.lazyMessage("altn8.pattern.postfixes").get()) {
+            routeToSettingPage(PostfixConfigurable::class.java)
         }
 
     private val freeRegexSettingsLink =
-        HyperlinkLabel(I18n.lazyMessage("altn8.freeRegexes").get()).apply {
-            addHyperlinkListener {
-                routeToSettingPage(FreeRegexConfigurable::class.java)
-            }
+        ActionLink(I18n.lazyMessage("altn8.freeRegexes").get()) {
+            routeToSettingPage(FreeRegexConfigurable::class.java)
         }
 
     override val panel: JPanel =
@@ -62,8 +55,6 @@ internal class GeneralLinksUi : Ui {
     override fun reset() {
     }
 
-    private fun routeToSettingPage(target: Class<out AbstractConfigurable>) {
-        ShowSettingsUtil.getInstance()
-            .showSettingsDialog(ProjectManager.getInstance().defaultProject, target)
-    }
+    private fun routeToSettingPage(target: Class<out AbstractConfigurable>) =
+        ShowSettingsUtil.getInstance().showSettingsDialog(null, target)
 }