+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Navigate to related files using a shortcut, which is (Alt)(8) by default.
 ### Changed
 
 * Use context help color for multiline descriptions
+* Add links to pattern settings
 
 ### Fixed